MLIR.AST.Dialect.Std

Documentation

pattern Branch :: Location -> BlockName -> [Name] -> Operation

br :: MonadBlockBuilder m => BlockName -> [Value] -> m EndOfBlock

cond_br :: MonadBlockBuilder m => Value -> BlockName -> [Value] -> BlockName -> [Value] -> m EndOfBlock

absf

The absf operation computes the absolute value. It takes one operand and returns one result of the same type. This type may be a float scalar type, a vector whose element type is float, or a tensor of floats.

Example:

// Scalar absolute value.
%a = absf %b : f64

// SIMD vector element-wise absolute value.
%f = absf %g : vector<4xf32>

// Tensor element-wise absolute value.
%x = absf %y : tensor<4x?xf8>

pattern AbsF :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.absf.

absf :: MonadBlockBuilder m => Value -> m Value

A builder for std.absf.

addf

Syntax:

operation ::= ssa-id = std.addf ssa-use , ssa-use : type

The addf operation takes two operands and returns one result, each of these is required to be the same type. This type may be a floating point scalar type, a vector whose element type is a floating point type, or a floating point tensor.

Example:

// Scalar addition.
%a = addf %b, %c : f64

// SIMD vector addition, e.g. for Intel SSE.
%f = addf %g, %h : vector<4xf32>

// Tensor addition.
%x = addf %y, %z : tensor<4x?xbf16>

TODO: In the distant future, this will accept optional attributes for fast math, contraction, rounding mode, and other controls.

pattern AddF :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.addf.

addf :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.addf.

addi

Syntax:

operation ::= ssa-id = std.addi ssa-use , ssa-use : type

The addi operation takes two operands and returns one result, each of these is required to be the same type. This type may be an integer scalar type, a vector whose element type is integer, or a tensor of integers. It has no standard attributes.

Example:

// Scalar addition.
%a = addi %b, %c : i64

// SIMD vector element-wise addition, e.g. for Intel SSE.
%f = addi %g, %h : vector<4xi32>

// Tensor element-wise addition.
%x = addi %y, %z : tensor<4x?xi8>

pattern AddI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.addi.

addi :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.addi.

and

Syntax:

operation ::= ssa-id = std.and ssa-use , ssa-use : type

The and operation takes two operands and returns one result, each of these is required to be the same type. This type may be an integer scalar type, a vector whose element type is integer, or a tensor of integers. It has no standard attributes.

Example:

// Scalar integer bitwise and.
%a = and %b, %c : i64

// SIMD vector element-wise bitwise integer and.
%f = and %g, %h : vector<4xi32>

// Tensor element-wise bitwise integer and.
%x = and %y, %z : tensor<4x?xi8>

pattern And :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.and.

and :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.and.

assert

Assert operation with single boolean operand and an error message attribute. If the argument is true this operation has no effect. Otherwise, the program execution will abort. The provided error message may be used by a runtime to propagate the error to the user.

Example:

assert %b, "Expected ... to be true"

pattern Assert :: Location -> operand -> ByteString -> AbstractOperation operand

A pattern for std.assert.

assert :: MonadBlockBuilder m => Value -> ByteString -> m ()

A builder for std.assert.

atomic_rmw

The atomic_rmw operation provides a way to perform a read-modify-write sequence that is free from data races. The kind enumeration specifies the modification to perform. The value operand represents the new value to be applied during the modification. The memref operand represents the buffer that the read and write will be performed against, as accessed by the specified indices. The arity of the indices is the rank of the memref. The result represents the latest value that was stored.

Example:

%x = atomic_rmw "addf" %value, %I[%i] : (f32, memref<10xf32>) -> f32

atomic_yield

"atomic_yield" yields an SSA value from a GenericAtomicRMWOp region.

pattern AtomicYield :: Location -> operand -> AbstractOperation operand

A pattern for std.atomic_yield.

atomic_yield :: MonadBlockBuilder m => Value -> m EndOfBlock

A builder for std.atomic_yield.

bitcast

Bitcast an integer or floating point value to an integer or floating point value of equal bit width. When operating on vectors, casts elementwise.

Note that this implements a logical bitcast independent of target endianness. This allows constant folding without target information and is consitent with the bitcast constant folders in LLVM (see https://github.com/llvm/llvm-project/blob/18c19414eb/llvm/lib/IR/ConstantFold.cpp\#L168) For targets where the source and target type have the same endianness (which is the standard), this cast will also change no bits at runtime, but it may still require an operation, for example if the machine has different floating point and integer register files. For targets that have a different endianness for the source and target types (e.g. float is big-endian and integer is little-endian) a proper lowering would add operations to swap the order of words in addition to the bitcast.

pattern Bitcast :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.bitcast.

bitcast :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.bitcast.

br

The br operation represents a branch operation in a function. The operation takes variable number of operands and produces no results. The operand number and types for each successor must match the arguments of the block successor.

Example:

^bb2:
  %2 = call @someFn()
  br ^bb3(%2 : tensor<*xf32>)
^bb3(%3: tensor<*xf32>):

call_indirect

The call_indirect operation represents an indirect call to a value of function type. Functions are first class types in MLIR, and may be passed as arguments and merged together with block arguments. The operands and result types of the call must match the specified function type.

Function values can be created with the constant operation.

Example:

%31 = call_indirect %15(%0, %1)
        : (tensor<16xf32>, tensor<16xf32>) -> tensor<16xf32>

call_indirect :: MonadBlockBuilder m => [Type] -> Value -> [Value] -> m Value

A builder for std.call_indirect.

call

The call operation represents a direct call to a function that is within the same symbol scope as the call. The operands and result types of the call must match the specified function type. The callee is encoded as a symbol reference attribute named "callee".

Example:

%2 = call @my_add(%0, %1) : (f32, f32) -> f32

ceilf

Syntax:

operation ::= ssa-id = std.ceilf ssa-use : type

The ceilf operation computes the ceiling of a given value. It takes one operand and returns one result of the same type. This type may be a float scalar type, a vector whose element type is float, or a tensor of floats. It has no standard attributes.

Example:

// Scalar ceiling value.
%a = ceilf %b : f64

// SIMD vector element-wise ceiling value.
%f = ceilf %g : vector<4xf32>

// Tensor element-wise ceiling value.
%x = ceilf %y : tensor<4x?xf8>

pattern CeilF :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.ceilf.

ceilf :: MonadBlockBuilder m => Value -> m Value

A builder for std.ceilf.

cmpf

The cmpf operation compares its two operands according to the float comparison rules and the predicate specified by the respective attribute. The predicate defines the type of comparison: (un)orderedness, (in)equality and signed less/greater than (or equal to) as well as predicates that are always true or false. The operands must have the same type, and this type must be a float type, or a vector or tensor thereof. The result is an i1, or a vector/tensor thereof having the same shape as the inputs. Unlike cmpi, the operands are always treated as signed. The u prefix indicates *unordered* comparison, not unsigned comparison, so "une" means unordered or not equal. For the sake of readability by humans, custom assembly form for the operation uses a string-typed attribute for the predicate. The value of this attribute corresponds to lower-cased name of the predicate constant, e.g., "one" means "ordered not equal". The string representation of the attribute is merely a syntactic sugar and is converted to an integer attribute by the parser.

Example:

%r1 = cmpf "oeq" %0, %1 : f32
%r2 = cmpf "ult" %0, %1 : tensor<42x42xf64>
%r3 = "std.cmpf"(%0, %1) {predicate: 0} : (f8, f8) -> i1

cmpi

The cmpi operation is a generic comparison for integer-like types. Its two arguments can be integers, vectors or tensors thereof as long as their types match. The operation produces an i1 for the former case, a vector or a tensor of i1 with the same shape as inputs in the other cases.

Its first argument is an attribute that defines which type of comparison is performed. The following comparisons are supported:

equal (mnemonic: "eq"; integer value: 0)

not equal (mnemonic: "ne"; integer value: 1)

signed less than (mnemonic: "slt"; integer value: 2)

signed less than or equal (mnemonic: "sle"; integer value: 3)

signed greater than (mnemonic: "sgt"; integer value: 4)

signed greater than or equal (mnemonic: "sge"; integer value: 5)

unsigned less than (mnemonic: "ult"; integer value: 6)

unsigned less than or equal (mnemonic: "ule"; integer value: 7)

unsigned greater than (mnemonic: "ugt"; integer value: 8)

unsigned greater than or equal (mnemonic: "uge"; integer value: 9)

The result is 1 if the comparison is true and 0 otherwise. For vector or tensor operands, the comparison is performed elementwise and the element of the result indicates whether the comparison is true for the operand elements with the same indices as those of the result.

Note: while the custom assembly form uses strings, the actual underlying attribute has integer type (or rather enum class in C++ code) as seen from the generic assembly form. String literals are used to improve readability of the IR by humans.

This operation only applies to integer-like operands, but not floats. The main reason being that comparison operations have diverging sets of attributes: integers require sign specification while floats require various floating point-related particularities, e.g., -ffast-math behavior, IEEE754 compliance, etc (rationale. The type of comparison is specified as attribute to avoid introducing ten similar operations, taking into account that they are often implemented using the same operation downstream (rationale. The separation between signed and unsigned order comparisons is necessary because of integers being signless. The comparison operation must know how to interpret values with the foremost bit being set: negatives in two's complement or large positives (rationale.

Example:

// Custom form of scalar "signed less than" comparison.
%x = cmpi "slt", %lhs, %rhs : i32

// Generic form of the same operation.
%x = "std.cmpi"(%lhs, %rhs) {predicate = 2 : i64} : (i32, i32) -> i1

// Custom form of vector equality comparison.
%x = cmpi "eq", %lhs, %rhs : vector<4xi64>

// Generic form of the same operation.
%x = "std.cmpi"(%lhs, %rhs) {predicate = 0 : i64}
    : (vector<4xi64>, vector<4xi64>) -> vector<4xi1>

cond_br

The cond_br terminator operation represents a conditional branch on a boolean (1-bit integer) value. If the bit is set, then the first destination is jumped to; if it is false, the second destination is chosen. The count and types of operands must align with the arguments in the corresponding target blocks.

The MLIR conditional branch operation is not allowed to target the entry block for a region. The two destinations of the conditional branch operation are allowed to be the same.

The following example illustrates a function with a conditional branch operation that targets the same block.

Example:

func @select(%a: i32, %b: i32, %flag: i1) -> i32 {
  // Both targets are the same, operands differ
  cond_br %flag, ^bb1(%a : i32), ^bb1(%b : i32)

^bb1(%x : i32) :
  return %x : i32
}

constant

Syntax:

operation ::= ssa-id = std.constant attribute-value : type

The constant operation produces an SSA value equal to some constant specified by an attribute. This is the way that MLIR uses to form simple integer and floating point constants, as well as more exotic things like references to functions and tensor/vector constants.

Example:

// Integer constant
%1 = constant 42 : i32

// Reference to function @myfn.
%3 = constant @myfn : (tensor<16xf32>, f32) -> tensor<16xf32>

// Equivalent generic forms
%1 = "std.constant"() {value = 42 : i32} : () -> i32
%3 = "std.constant"() {value = @myfn}
   : () -> ((tensor<16xf32>, f32) -> tensor<16xf32>)

MLIR does not allow direct references to functions in SSA operands because the compiler is multithreaded, and disallowing SSA values to directly reference a function simplifies this (rationale.

pattern Constant :: Location -> Type -> Attribute -> AbstractOperation operand

A pattern for std.constant.

constant :: MonadBlockBuilder m => Type -> Attribute -> m Value

A builder for std.constant.

copysign

Syntax:

operation ::= ssa-id = std.copysign ssa-use , ssa-use : type

The copysign returns a value with the magnitude of the first operand and the sign of the second operand. It takes two operands and returns one result of the same type. This type may be a float scalar type, a vector whose element type is float, or a tensor of floats. It has no standard attributes.

Example:

// Scalar copysign value.
%a = copysign %b, %c : f64

// SIMD vector element-wise copysign value.
%f = copysign %g, %h : vector<4xf32>

// Tensor element-wise copysign value.
%x = copysign %y, %z : tensor<4x?xf8>

pattern CopySign :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.copysign.

copysign :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.copysign.

divf

pattern DivF :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.divf.

divf :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.divf.

fpext

Cast a floating-point value to a larger floating-point-typed value. The destination type must to be strictly wider than the source type. When operating on vectors, casts elementwise.

pattern FPExt :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.fpext.

fpext :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.fpext.

fptosi

Cast from a value interpreted as floating-point to the nearest (rounding towards zero) signed integer value. When operating on vectors, casts elementwise.

pattern FPToSI :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.fptosi.

fptosi :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.fptosi.

fptoui

Cast from a value interpreted as floating-point to the nearest (rounding towards zero) unsigned integer value. When operating on vectors, casts elementwise.

pattern FPToUI :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.fptoui.

fptoui :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.fptoui.

fptrunc

Truncate a floating-point value to a smaller floating-point-typed value. The destination type must be strictly narrower than the source type. If the value cannot be exactly represented, it is rounded using the default rounding mode. When operating on vectors, casts elementwise.

pattern FPTrunc :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.fptrunc.

fptrunc :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.fptrunc.

floorf

Syntax:

operation ::= ssa-id = std.floorf ssa-use : type

The floorf operation computes the floor of a given value. It takes one operand and returns one result of the same type. This type may be a float scalar type, a vector whose element type is float, or a tensor of floats. It has no standard attributes.

Example:

// Scalar floor value.
%a = floorf %b : f64

// SIMD vector element-wise floor value.
%f = floorf %g : vector<4xf32>

// Tensor element-wise floor value.
%x = floorf %y : tensor<4x?xf8>

pattern FloorF :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.floorf.

floorf :: MonadBlockBuilder m => Value -> m Value

A builder for std.floorf.

fmaf

Syntax:

operation ::= ssa-id = std.fmaf ssa-use , ssa-use , ssa-use : type

The fmaf operation takes three operands and returns one result, each of these is required to be the same type. This type may be a floating point scalar type, a vector whose element type is a floating point type, or a floating point tensor.

Example:

// Scalar fused multiply-add: d = a*b + c
%d = fmaf %a, %b, %c : f64

// SIMD vector fused multiply-add, e.g. for Intel SSE.
%i = fmaf %f, %g, %h : vector<4xf32>

// Tensor fused multiply-add.
%w = fmaf %x, %y, %z : tensor<4x?xbf16>

The semantics of the operation correspond to those of the llvm.fma intrinsic. In the particular case of lowering to LLVM, this is guaranteed to lower to the llvm.fma.* intrinsic.

pattern FmaF :: Location -> Type -> operand -> operand -> operand -> AbstractOperation operand

A pattern for std.fmaf.

fmaf :: MonadBlockBuilder m => Value -> Value -> Value -> m Value

A builder for std.fmaf.

generic_atomic_rmw

The generic_atomic_rmw operation provides a way to perform a read-modify-write sequence that is free from data races. The memref operand represents the buffer that the read and write will be performed against, as accessed by the specified indices. The arity of the indices is the rank of the memref. The result represents the latest value that was stored. The region contains the code for the modification itself. The entry block has a single argument that represents the value stored in memref[indices] before the write is performed. No side-effecting ops are allowed in the body of GenericAtomicRMWOp.

Example:

%x = generic_atomic_rmw %I[%i] : memref<10xf32> {
  ^bb0(%current_value : f32):
    %c1 = constant 1.0 : f32
    %inc = addf %c1, %current_value : f32
    atomic_yield %inc : f32
}

generic_atomic_rmw :: MonadBlockBuilder m => Type -> Value -> [Value] -> RegionBuilderT m () -> m Value

A builder for std.generic_atomic_rmw.

index_cast

Casts between scalar or vector integers and corresponding 'index' scalar or vectors. Index is an integer of platform-specific bit width. If casting to a wider integer, the value is sign-extended. If casting to a narrower integer, the value is truncated.

pattern IndexCast :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.index_cast.

index_cast :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.index_cast.

maxf

Syntax:

operation ::= ssa-id = maxf ssa-use , ssa-use : type

Returns the maximum of the two arguments, treating -0.0 as less than +0.0. If one of the arguments is NaN, then the result is also NaN.

Example:

// Scalar floating-point maximum.
%a = maxf %b, %c : f64

pattern MaxF :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.maxf.

maxf :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.maxf.

maxsi

Syntax:

operation ::= ssa-id = maxsi ssa-use , ssa-use : type

Returns the larger of %a and %b comparing the values as signed integers.

Example:

// Scalar signed integer maximum.
%a = maxsi %b, %c : i64

pattern MaxSI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.maxsi.

maxsi :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.maxsi.

maxui

Syntax:

operation ::= ssa-id = maxui ssa-use , ssa-use : type

Returns the larger of %a and %b comparing the values as unsigned integers.

Example:

// Scalar unsigned integer maximum.
%a = maxui %b, %c : i64

pattern MaxUI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.maxui.

maxui :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.maxui.

minf

Syntax:

operation ::= ssa-id = minf ssa-use , ssa-use : type

Returns the minimum of the two arguments, treating -0.0 as less than +0.0. If one of the arguments is NaN, then the result is also NaN.

Example:

// Scalar floating-point minimum.
%a = minf %b, %c : f64

pattern MinF :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.minf.

minf :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.minf.

minsi

Syntax:

operation ::= ssa-id = minsi ssa-use , ssa-use : type

Returns the smaller of %a and %b comparing the values as signed integers.

Example:

// Scalar signed integer minimum.
%a = minsi %b, %c : i64

pattern MinSI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.minsi.

minsi :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.minsi.

minui

Syntax:

operation ::= ssa-id = minui ssa-use , ssa-use : type

Returns the smaller of %a and %b comparing the values as unsigned integers.

Example:

// Scalar unsigned integer minimum.
%a = minui %b, %c : i64

pattern MinUI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.minui.

minui :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.minui.

mulf

Syntax:

operation ::= ssa-id = std.mulf ssa-use , ssa-use : type

The mulf operation takes two operands and returns one result, each of these is required to be the same type. This type may be a floating point scalar type, a vector whose element type is a floating point type, or a floating point tensor.

Example:

// Scalar multiplication.
%a = mulf %b, %c : f64

// SIMD pointwise vector multiplication, e.g. for Intel SSE.
%f = mulf %g, %h : vector<4xf32>

// Tensor pointwise multiplication.
%x = mulf %y, %z : tensor<4x?xbf16>

TODO: In the distant future, this will accept optional attributes for fast math, contraction, rounding mode, and other controls.

pattern MulF :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.mulf.

mulf :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.mulf.

muli

pattern MulI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.muli.

muli :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.muli.

negf

Syntax:

operation ::= ssa-id = negf ssa-use : type

The negf operation computes the negation of a given value. It takes one operand and returns one result of the same type. This type may be a float scalar type, a vector whose element type is float, or a tensor of floats. It has no standard attributes.

Example:

// Scalar negation value.
%a = negf %b : f64

// SIMD vector element-wise negation value.
%f = negf %g : vector<4xf32>

// Tensor element-wise negation value.
%x = negf %y : tensor<4x?xf8>

pattern NegF :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.negf.

negf :: MonadBlockBuilder m => Value -> m Value

A builder for std.negf.

or

Syntax:

operation ::= ssa-id = or ssa-use , ssa-use : type

The or operation takes two operands and returns one result, each of these is required to be the same type. This type may be an integer scalar type, a vector whose element type is integer, or a tensor of integers. It has no standard attributes.

Example:

// Scalar integer bitwise or.
%a = or %b, %c : i64

// SIMD vector element-wise bitwise integer or.
%f = or %g, %h : vector<4xi32>

// Tensor element-wise bitwise integer or.
%x = or %y, %z : tensor<4x?xi8>

pattern Or :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.or.

or :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.or.

rank

The rank operation takes a memref/tensor operand and returns its rank.

Example:

%1 = rank %arg0 : tensor<*xf32>
%2 = rank %arg1 : memref<*xf32>

pattern Rank :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.rank.

rank :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.rank.

remf

pattern RemF :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.remf.

remf :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.remf.

return

The return operation represents a return operation within a function. The operation takes variable number of operands and produces no results. The operand number and types must match the signature of the function that contains the operation.

Example:

func @foo() : (i32, f8) {
  ...
  return %0, %1 : i32, f8
}

pattern Return :: Location -> [operand] -> AbstractOperation operand

A pattern for std.return.

return :: MonadBlockBuilder m => [Value] -> m EndOfBlock

A builder for std.return.

sitofp

Cast from a value interpreted as a signed integer to the corresponding floating-point value. If the value cannot be exactly represented, it is rounded using the default rounding mode. When operating on vectors, casts elementwise.

pattern SIToFP :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.sitofp.

sitofp :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.sitofp.

select

The select operation chooses one value based on a binary condition supplied as its first operand. If the value of the first operand is 1, the second operand is chosen, otherwise the third operand is chosen. The second and the third operand must have the same type.

The operation applies to vectors and tensors elementwise given the _shape_ of all operands is identical. The choice is made for each element individually based on the value at the same position as the element in the condition operand. If an i1 is provided as the condition, the entire vector or tensor is chosen.

The select operation combined with cmpi can be used to implement min and max with signed or unsigned comparison semantics.

Example:

// Custom form of scalar selection.
%x = select %cond, %true, %false : i32

// Generic form of the same operation.
%x = "std.select"(%cond, %true, %false) : (i1, i32, i32) -> i32

// Element-wise vector selection.
%vx = std.select %vcond, %vtrue, %vfalse : vector<42xi1>, vector<42xf32>

// Full vector selection.
%vx = std.select %cond, %vtrue, %vfalse : vector<42xf32>

pattern Select :: Location -> Type -> operand -> operand -> operand -> AbstractOperation operand

A pattern for std.select.

select :: MonadBlockBuilder m => Type -> Value -> Value -> Value -> m Value

A builder for std.select.

shift_left

The shift_left operation shifts an integer value to the left by a variable amount. The low order bits are filled with zeros.

Example:

%1 = constant 5 : i8                       // %1 is 0b00000101
%2 = constant 3 : i8
%3 = shift_left %1, %2 : (i8, i8) -> i8    // %3 is 0b00101000

pattern ShiftLeft :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.shift_left.

shift_left :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.shift_left.

sexti

The integer sign extension operation takes an integer input of width M and an integer destination type of width N. The destination bit-width must be larger than the input bit-width (N > M). The top-most (N - M) bits of the output are filled with copies of the most-significant bit of the input.

Example:

%1 = constant 5 : i3            // %1 is 0b101
%2 = sexti %1 : i3 to i6        // %2 is 0b111101
%3 = constant 2 : i3            // %3 is 0b010
%4 = sexti %3 : i3 to i6        // %4 is 0b000010

%5 = sexti %0 : vector<2 x i32> to vector<2 x i64>

pattern SignExtendI :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.sexti.

sexti :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.sexti.

ceildivi_signed

Syntax:

operation ::= ssa-id = ceildivi_signed ssa-use , ssa-use : type

Signed integer division. Rounds towards positive infinity, i.e. 7 / -2 = -3.

Note: the semantics of division by zero or signed division overflow (minimum value divided by -1) is TBD; do NOT assume any specific behavior.

Example:

// Scalar signed integer division.
%a = ceildivi_signed %b, %c : i64

pattern SignedCeilDivI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.ceildivi_signed.

ceildivi_signed :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.ceildivi_signed.

divi_signed

Syntax:

operation ::= ssa-id = divi_signed ssa-use , ssa-use : type

Signed integer division. Rounds towards zero. Treats the leading bit as sign, i.e. 6 / -2 = -3.

Note: the semantics of division by zero or signed division overflow (minimum value divided by -1) is TBD; do NOT assume any specific behavior.

Example:

// Scalar signed integer division.
%a = divi_signed %b, %c : i64

// SIMD vector element-wise division.
%f = divi_signed %g, %h : vector<4xi32>

// Tensor element-wise integer division.
%x = divi_signed %y, %z : tensor<4x?xi8>

pattern SignedDivI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.divi_signed.

divi_signed :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.divi_signed.

floordivi_signed

Syntax:

operation ::= ssa-id = floordivi_signed ssa-use , ssa-use : type

Signed integer division. Rounds towards negative infinity, i.e. 5 / -2 = -3.

Note: the semantics of division by zero or signed division overflow (minimum value divided by -1) is TBD; do NOT assume any specific behavior.

Example:

// Scalar signed integer division.
%a = floordivi_signed %b, %c : i64

pattern SignedFloorDivI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.floordivi_signed.

floordivi_signed :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.floordivi_signed.

remi_signed

Syntax:

operation ::= ssa-id = std.remi_signed ssa-use , ssa-use : type

Signed integer division remainder. Treats the leading bit as sign, i.e. 6 % -2 = 0.

Note: the semantics of division by zero is TBD; do NOT assume any specific behavior.

Example:

// Scalar signed integer division remainder.
%a = remi_signed %b, %c : i64

// SIMD vector element-wise division remainder.
%f = remi_signed %g, %h : vector<4xi32>

// Tensor element-wise integer division remainder.
%x = remi_signed %y, %z : tensor<4x?xi8>

pattern SignedRemI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.remi_signed.

remi_signed :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.remi_signed.

shift_right_signed

The shift_right_signed operation shifts an integer value to the right by a variable amount. The integer is interpreted as signed. The high order bits in the output are filled with copies of the most-significant bit of the shifted value (which means that the sign of the value is preserved).

Example:

%1 = constant 160 : i8                             // %1 is 0b10100000
%2 = constant 3 : i8
%3 = shift_right_signed %1, %2 : (i8, i8) -> i8    // %3 is 0b11110100
%4 = constant 96 : i8                              // %4 is 0b01100000
%5 = shift_right_signed %4, %2 : (i8, i8) -> i8    // %5 is 0b00001100

pattern SignedShiftRight :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.shift_right_signed.

shift_right_signed :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.shift_right_signed.

splat

Broadcast the operand to all elements of the result vector or tensor. The operand has to be of integerindexfloat type. When the result is a tensor, it has to be statically shaped.

Example:

%s = load %A[%i] : memref<128xf32>
%v = splat %s : vector<4xf32>
%t = splat %s : tensor<8x16xi32>

TODO: This operation is easy to extend to broadcast to dynamically shaped tensors in the same way dynamically shaped memrefs are handled.

// Broadcasts %s to a 2-d dynamically shaped tensor, with %m, %n binding
// to the sizes of the two dynamic dimensions.
%m = "foo"() : () -> (index)
%n = "bar"() : () -> (index)
%t = splat %s [%m, %n] : tensor<?x?xi32>

pattern Splat :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.splat.

splat :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.splat.

subf

pattern SubF :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.subf.

subf :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.subf.

subi

pattern SubI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.subi.

subi :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.subi.

switch

The switch terminator operation represents a switch on a signless integer value. If the flag matches one of the specified cases, then the corresponding destination is jumped to. If the flag does not match any of the cases, the default destination is jumped to. The count and types of operands must align with the arguments in the corresponding target blocks.

Example:

switch %flag : i32, [
  default: ^bb1(%a : i32),
  42: ^bb1(%b : i32),
  43: ^bb3(%c : i32)
]

trunci

The integer truncation operation takes an integer input of width M and an integer destination type of width N. The destination bit-width must be smaller than the input bit-width (N < M). The top-most (N - M) bits of the input are discarded.

Example:

  %1 = constant 21 : i5           // %1 is 0b10101
  %2 = trunci %1 : i5 to i4       // %2 is 0b0101
  %3 = trunci %1 : i5 to i3       // %3 is 0b101

  %5 = trunci %0 : vector<2 x i32> to vector<2 x i16>

pattern TruncateI :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.trunci.

trunci :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.trunci.

uitofp

Cast from a value interpreted as unsigned integer to the corresponding floating-point value. If the value cannot be exactly represented, it is rounded using the default rounding mode. When operating on vectors, casts elementwise.

pattern UIToFP :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.uitofp.

uitofp :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.uitofp.

divi_unsigned

Syntax: operation ::= ssa-id = std.divi_unsigned ssa-use , ssa-use : type

Unsigned integer division. Rounds towards zero. Treats the leading bit as the most significant, i.e. for i16 given two's complement representation, 6 -2 = 6 (2^16 - 2) = 0.

Note: the semantics of division by zero is TBD; do NOT assume any specific behavior.

Example:

// Scalar unsigned integer division.
%a = divi_unsigned %b, %c : i64

// SIMD vector element-wise division.
%f = divi_unsigned %g, %h : vector<4xi32>

// Tensor element-wise integer division.
%x = divi_unsigned %y, %z : tensor<4x?xi8>

pattern UnsignedDivI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.divi_unsigned.

divi_unsigned :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.divi_unsigned.

remi_unsigned

Syntax:

operation ::= ssa-id = std.remi_unsigned ssa-use , ssa-use : type

Unsigned integer division remainder. Treats the leading bit as the most significant, i.e. for i16, 6 % -2 = 6 % (2^16 - 2) = 6.

Note: the semantics of division by zero is TBD; do NOT assume any specific behavior.

Example:

// Scalar unsigned integer division remainder.
%a = remi_unsigned %b, %c : i64

// SIMD vector element-wise division remainder.
%f = remi_unsigned %g, %h : vector<4xi32>

// Tensor element-wise integer division remainder.
%x = remi_unsigned %y, %z : tensor<4x?xi8>

pattern UnsignedRemI :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.remi_unsigned.

remi_unsigned :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.remi_unsigned.

shift_right_unsigned

The shift_right_unsigned operation shifts an integer value to the right by a variable amount. The integer is interpreted as unsigned. The high order bits are always filled with zeros.

Example:

%1 = constant 160 : i8                               // %1 is 0b10100000
%2 = constant 3 : i8
%3 = shift_right_unsigned %1, %2 : (i8, i8) -> i8    // %3 is 0b00010100

pattern UnsignedShiftRight :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.shift_right_unsigned.

shift_right_unsigned :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.shift_right_unsigned.

xor

The xor operation takes two operands and returns one result, each of these is required to be the same type. This type may be an integer scalar type, a vector whose element type is integer, or a tensor of integers. It has no standard attributes.

Example:

// Scalar integer bitwise xor.
%a = xor %b, %c : i64

// SIMD vector element-wise bitwise integer xor.
%f = xor %g, %h : vector<4xi32>

// Tensor element-wise bitwise integer xor.
%x = xor %y, %z : tensor<4x?xi8>

pattern XOr :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for std.xor.

xor :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for std.xor.

zexti

The integer zero extension operation takes an integer input of width M and an integer destination type of width N. The destination bit-width must be larger than the input bit-width (N > M). The top-most (N - M) bits of the output are filled with zeros.

Example:

  %1 = constant 5 : i3            // %1 is 0b101
  %2 = zexti %1 : i3 to i6        // %2 is 0b000101
  %3 = constant 2 : i3            // %3 is 0b010
  %4 = zexti %3 : i3 to i6        // %4 is 0b000010

  %5 = zexti %0 : vector<2 x i32> to vector<2 x i64>

pattern ZeroExtendI :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.zexti.

zexti :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.zexti.