MLIR.Native

Description

This module defines a set of Haskell types wrapping references to native C++ MLIR objects along with some basic operations on them. See the submodules for more specialized components such as an ExecutionEngine or PassManager.

Contexts

data Context

A native MLIR context.

createContext :: IO Context

Creates a native MLIR context.

destroyContext :: Context -> IO ()

Destroys a native MLIR context.

withContext :: (Context -> IO a) -> IO a

Wraps an IO action that gets access to a fresh MLIR context.

class HasContext a where

A typeclass for retrieving MLIR contexts managing other native types.

Methods

getContext :: a -> IO Context

Retrieve the MLIR context that manages the storage of the native value.

Dialect registration

registerAllDialects :: Context -> IO ()

Register all builtin MLIR dialects in the specified Context.

getNumLoadedDialects :: Context -> IO Int

Retrieve the count of dialects currently registered in the Context.

Type

data Type

A native MLIR type object.

Location

data Location

A native MLIR location object.

getFileLineColLocation :: Context -> StringRef -> CUInt -> CUInt -> IO Location

getNameLocation :: Context -> StringRef -> Location -> IO Location

getUnknownLocation :: Context -> IO Location

Create an unknown source location.

Operation

data Operation

A native MLIR operation instance.

getOperationName :: Operation -> IO Identifier

Retrieve the name of the given operation.

showOperation :: Operation -> IO ByteString

Show the operation using the MLIR printer.

showOperationWithLocation :: Operation -> IO ByteString

Show the operation with location using the MLIR printer.

verifyOperation :: Operation -> IO Bool

Check validity of the operation.

Region

data Region

A native MLIR region.

getOperationRegion :: CIntPtr -> Operation -> IO Region

Returns the pos'th region of an Operation.

getRegionBlocks :: Region -> IO [Block]

Returns the Blocks in a Region.

Block

data Block

A native MLIR block object. Every block is a list of Operations.

showBlock :: Block -> IO ByteString

Show the block using the MLIR printer.

getBlockOperations :: Block -> IO [Operation]

Returns the Operations in a Block.

Module

data Module

A native MLIR module operation. Since every module is an operation, it can be converted to an Operation using moduleAsOperation.

createEmptyModule :: Location -> IO Module

Create an empty module.

parseModule :: Context -> StringRef -> IO (Maybe Module)

Parse a module from a string. Returns Nothing in case of parse failure.

destroyModule :: Module -> IO ()

Destroy all resources associated with a Module.

getModuleBody :: Module -> IO Block

Retrieve the block containg all module definitions.

moduleAsOperation :: Module -> IO Operation

Convert a module to an Operation.

moduleFromOperation :: Operation -> IO (Maybe Module)

Inverse of moduleAsOperation. Returns Nothing if the operation is not a builtin MLIR module operation.

showModule :: Module -> IO ByteString

Show the module using the MLIR printer.

StringRef

data StringRef

Constructors

StringRef (Ptr CChar) CSize

withStringRef :: ByteString -> (StringRef -> IO a) -> IO a

Use a ByteString as a StringRef. This is O(n) due to MLIR sometimes requiring the StringRefs to be null-terminated.

Identifier

data Identifier

A native MLIR identifier. Identifiers are strings interned in the MLIR context.

createIdentifier :: Context -> StringRef -> IO Identifier

Create an identifier from a StringRef.

identifierString :: Identifier -> IO StringRef

View an identifier as a StringRef. The result is valid for as long as the Context managing the identifier.

LogicalResult

data LogicalResult

A result code for many failable MLIR operations. The only valid cases are Success and Failure.

pattern Failure :: LogicalResult

Indicates a filure of an MLIR operation. Inspect the diagnostics output to find the cause of the issue.

pattern Success :: LogicalResult

Indicates a successful completion of an MLIR operation.

Debugging utilities

setDebugMode :: Bool -> IO ()

Enable or disable debug logging in MLIR.

class HasDump a where

A class for native objects that can be dumped to standard error output.

Methods

dump :: a -> IO ()

Display the value in the standard error output.