MLIR.AST.Dialect.LLVM

Types

data Type

Constructors

PointerType Type

ArrayType Int Type

VoidType

LiteralStructType [Type]

pattern Ptr :: Type -> Type

pattern Array :: Int -> Type -> Type

pattern Void :: Type

pattern LiteralStruct :: [Type] -> Type

Operations

ashr

pattern AShr :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.ashr.

ashr :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.ashr.

access_group

Defines an access group metadata that can be attached to any instruction that potentially accesses memory. The access group may be attached to a memory accessing instruction via the llvm.access.group metadata and a branch instruction in the loop latch block via the llvm.loop.parallel_accesses metadata.

See the following link for more details: https://llvm.org/docs/LangRef.html\#llvm-access-group-metadata

add

pattern Add :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.add.

add :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.add.

addrspacecast

pattern AddrSpaceCast :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.addrspacecast.

addrspacecast :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.addrspacecast.

mlir.addressof

Creates an SSA value containing a pointer to a global variable or constant defined by llvm.mlir.global. The global value can be defined after its first referenced. If the global value is a constant, storing into it is not allowed.

Examples:

func @foo() {
  // Get the address of a global variable.
  %0 = llvm.mlir.addressof @const : !llvm.ptr<i32>

  // Use it as a regular pointer.
  %1 = llvm.load %0 : !llvm.ptr<i32>

  // Get the address of a function.
  %2 = llvm.mlir.addressof @foo : !llvm.ptr<func<void ()>>

  // The function address can be used for indirect calls.
  llvm.call %2() : () -> ()
}

// Define the global.
llvm.mlir.global @const(42 : i32) : i32

alias_scope_domain

Defines a domain that may be associated with an alias scope.

See the following link for more details: https://llvm.org/docs/LangRef.html\#noalias-and-alias-scope-metadata

alias_scope

Defines an alias scope that can be attached to a memory-accessing operation. Such scopes can be used in combination with noalias metadata to indicate that sets of memory-affecting operations in one scope do not alias with memory-affecting operations in another scope.

Example: module { llvm.func @foo(%ptr1 : !llvm.ptr<i32>) { %c0 = llvm.mlir.constant(0 : i32) : i32 %c4 = llvm.mlir.constant(4 : i32) : i32 %1 = llvm.ptrtoint %ptr1 : !llvm.ptr<i32> to i32 %2 = llvm.add %1, %c1 : i32 %ptr2 = llvm.inttoptr %2 : i32 to !llvm.ptr<i32> llvm.store %c0, %ptr1 { alias_scopes = [@metadata::@scope1], llvm.noalias = [@metadata::@scope2] } : !llvm.ptr<i32> llvm.store %c4, %ptr2 { alias_scopes = [@metadata::@scope2], llvm.noalias = [@metadata::@scope1] } : !llvm.ptr<i32> llvm.return }

llvm.metadata @metadata { llvm.alias_scope_domain @unused_domain llvm.alias_scope_domain @domain { description = "Optional domain description"} llvm.alias_scope @scope1 { domain = @domain } llvm.alias_scope @scope2 { domain = @domain, description = "Optional scope description" } llvm.return } }

See the following link for more details: https://llvm.org/docs/LangRef.html\#noalias-and-alias-scope-metadata

alloca

pattern Alloca :: Location -> Type -> operand -> Maybe Int -> AbstractOperation operand

A pattern for llvm.alloca.

alloca :: MonadBlockBuilder m => Type -> Value -> Maybe Int -> m Value

A builder for llvm.alloca.

and

pattern And :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.and.

and :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.and.

intr.assume

pattern Assume :: Location -> operand -> AbstractOperation operand

A pattern for llvm.intr.assume.

intr_assume :: MonadBlockBuilder m => Value -> m ()

A builder for llvm.intr.assume.

cmpxchg

atomicrmw

intr.bitreverse

pattern BitReverse :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.bitreverse.

intr_bitreverse :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.bitreverse.

bitcast

pattern Bitcast :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.bitcast.

bitcast :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.bitcast.

br

call

In LLVM IR, functions may return either 0 or 1 value. LLVM IR dialect implements this behavior by providing a variadic call operation for 0- and 1-result functions. Even though MLIR supports multi-result functions, LLVM IR dialect disallows them.

The call instruction supports both direct and indirect calls. Direct calls start with a function name (@-prefixed) and indirect calls start with an SSA value (%-prefixed). The direct callee, if present, is stored as a function attribute callee. The trailing type of the instruction is always the MLIR function type, which may be different from the indirect callee that has the wrapped LLVM IR function type.

Examples:

// Direct call without arguments and with one result.
%0 = llvm.call @foo() : () -> (f32)

// Direct call with arguments and without a result.
llvm.call @bar(%0) : (f32) -> ()

// Indirect call with an argument and without a result.
llvm.call %1(%0) : (f32) -> ()

call :: MonadBlockBuilder m => [Type] -> [Value] -> m Value

A builder for llvm.call.

cond_br

mlir.constant

Unlike LLVM IR, MLIR does not have first-class constant values. Therefore, all constants must be created as SSA values before being used in other operations. llvm.mlir.constant creates such values for scalars and vectors. It has a mandatory value attribute, which may be an integer, floating point attribute; dense or sparse attribute containing integers or floats. The type of the attribute is one of the corresponding MLIR builtin types. It may be omitted for i64 and f64 types that are implied. The operation produces a new SSA value of the specified LLVM IR dialect type. The type of that value _must_ correspond to the attribute type converted to LLVM IR.

Examples:

// Integer constant, internal i32 is mandatory
%0 = llvm.mlir.constant(42 : i32) : i32

// It's okay to omit i64.
%1 = llvm.mlir.constant(42) : i64

// Floating point constant.
%2 = llvm.mlir.constant(42.0 : f32) : f32

// Splat dense vector constant.
%3 = llvm.mlir.constant(dense<1.0> : vector<4xf32>) : vector<4xf32>

pattern Constant :: Location -> Type -> Attribute -> AbstractOperation operand

A pattern for llvm.mlir.constant.

mlir_constant :: MonadBlockBuilder m => Type -> Attribute -> m Value

A builder for llvm.mlir.constant.

intr.copysign

pattern CopySign :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.copysign.

intr_copysign :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.intr.copysign.

intr.coro.begin

pattern CoroBegin :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.coro.begin.

intr_coro_begin :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.intr.coro.begin.

intr.coro.end

pattern CoroEnd :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.coro.end.

intr_coro_end :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.intr.coro.end.

intr.coro.free

pattern CoroFree :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.coro.free.

intr_coro_free :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.intr.coro.free.

intr.coro.id

pattern CoroId :: Location -> Type -> operand -> operand -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.coro.id.

intr_coro_id :: MonadBlockBuilder m => Type -> Value -> Value -> Value -> Value -> m Value

A builder for llvm.intr.coro.id.

intr.coro.resume

pattern CoroResume :: Location -> operand -> AbstractOperation operand

A pattern for llvm.intr.coro.resume.

intr_coro_resume :: MonadBlockBuilder m => Value -> m ()

A builder for llvm.intr.coro.resume.

intr.coro.save

pattern CoroSave :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.coro.save.

intr_coro_save :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.coro.save.

intr.coro.size

pattern CoroSize :: Location -> Type -> AbstractOperation operand

A pattern for llvm.intr.coro.size.

intr_coro_size :: MonadBlockBuilder m => Type -> m Value

A builder for llvm.intr.coro.size.

intr.coro.suspend

pattern CoroSuspend :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.coro.suspend.

intr_coro_suspend :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.intr.coro.suspend.

intr.cos

pattern Cos :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.cos.

intr_cos :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.cos.

intr.ctpop

pattern CtPop :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.ctpop.

intr_ctpop :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.ctpop.

intr.exp2

pattern Exp2 :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.exp2.

intr_exp2 :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.exp2.

intr.exp

pattern Exp :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.exp.

intr_exp :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.exp.

extractelement

pattern ExtractElement :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.extractelement.

extractelement :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.extractelement.

extractvalue

pattern ExtractValue :: Location -> Type -> operand -> [Attribute] -> AbstractOperation operand

A pattern for llvm.extractvalue.

extractvalue :: MonadBlockBuilder m => Type -> Value -> [Attribute] -> m Value

A builder for llvm.extractvalue.

intr.fabs

pattern FAbs :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.fabs.

intr_fabs :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.fabs.

fadd

pattern FAdd :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.fadd.

fadd :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.fadd.

intr.ceil

pattern FCeil :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.ceil.

intr_ceil :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.ceil.

fcmp

fdiv

pattern FDiv :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.fdiv.

fdiv :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.fdiv.

intr.floor

pattern FFloor :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.floor.

intr_floor :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.floor.

intr.fma

pattern FMA :: Location -> Type -> operand -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.fma.

intr_fma :: MonadBlockBuilder m => Value -> Value -> Value -> m Value

A builder for llvm.intr.fma.

intr.fmuladd

pattern FMulAdd :: Location -> Type -> operand -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.fmuladd.

intr_fmuladd :: MonadBlockBuilder m => Value -> Value -> Value -> m Value

A builder for llvm.intr.fmuladd.

fmul

pattern FMul :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.fmul.

fmul :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.fmul.

fneg

pattern FNeg :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.fneg.

fneg :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.fneg.

fpext

pattern FPExt :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.fpext.

fpext :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.fpext.

fptosi

pattern FPToSI :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.fptosi.

fptosi :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.fptosi.

fptoui

pattern FPToUI :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.fptoui.

fptoui :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.fptoui.

fptrunc

pattern FPTrunc :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.fptrunc.

fptrunc :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.fptrunc.

frem

pattern FRem :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.frem.

frem :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.frem.

fsub

pattern FSub :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.fsub.

fsub :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.fsub.

fence

freeze

pattern Freeze :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.freeze.

freeze :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.freeze.

getelementptr

getelementptr :: MonadBlockBuilder m => Type -> Value -> [Value] -> m Value

A builder for llvm.getelementptr.

intr.get.active.lane.mask

pattern GetActiveLaneMask :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.get.active.lane.mask.

intr_get_active_lane_mask :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.intr.get.active.lane.mask.

mlir.global

Since MLIR allows for arbitrary operations to be present at the top level, global variables are defined using the llvm.mlir.global operation. Both global constants and variables can be defined, and the value may also be initialized in both cases.

There are two forms of initialization syntax. Simple constants that can be represented as MLIR attributes can be given in-line:

llvm.mlir.global @variable(32.0 : f32) : f32

This initialization and type syntax is similar to llvm.mlir.constant and may use two types: one for MLIR attribute and another for the LLVM value. These types must be compatible.

More complex constants that cannot be represented as MLIR attributes can be given in an initializer region:

// This global is initialized with the equivalent of:
//   i32* getelementptr (i32* @g2, i32 2)
llvm.mlir.global constant @int_gep() : !llvm.ptr<i32> {
  %0 = llvm.mlir.addressof @g2 : !llvm.ptr<i32>
  %1 = llvm.mlir.constant(2 : i32) : i32
  %2 = llvm.getelementptr %0[%1]
     : (!llvm.ptr<i32>, i32) -> !llvm.ptr<i32>
  // The initializer region must end with llvm.return.
  llvm.return %2 : !llvm.ptr<i32>
}

Only one of the initializer attribute or initializer region may be provided.

llvm.mlir.global must appear at top-level of the enclosing module. It uses an @-identifier for its value, which will be uniqued by the module with respect to other @-identifiers in it.

Examples:

// Global values use @-identifiers.
llvm.mlir.global constant @cst(42 : i32) : i32

// Non-constant values must also be initialized.
llvm.mlir.global @variable(32.0 : f32) : f32

// Strings are expected to be of wrapped LLVM i8 array type and do not
// automatically include the trailing zero.
llvm.mlir.global @string("abc") : !llvm.array<3 x i8>

// For strings globals, the trailing type may be omitted.
llvm.mlir.global constant @no_trailing_type("foo bar")

// A complex initializer is constructed with an initializer region.
llvm.mlir.global constant @int_gep() : !llvm.ptr<i32> {
  %0 = llvm.mlir.addressof @g2 : !llvm.ptr<i32>
  %1 = llvm.mlir.constant(2 : i32) : i32
  %2 = llvm.getelementptr %0[%1]
     : (!llvm.ptr<i32>, i32) -> !llvm.ptr<i32>
  llvm.return %2 : !llvm.ptr<i32>
}

Similarly to functions, globals have a linkage attribute. In the custom syntax, this attribute is placed between llvm.mlir.global and the optional constant keyword. If the attribute is omitted, external linkage is assumed by default.

Examples:

// A constant with internal linkage will not participate in linking.
llvm.mlir.global internal constant @cst(42 : i32) : i32

// By default, "external" linkage is assumed and the global participates in
// symbol resolution at link-time.
llvm.mlir.global @glob(0 : f32) : f32

// Alignment is optional
llvm.mlir.global private constant @y(dense<1.0> : tensor<8xf32>) : !llvm.array<8 x f32>

Like global variables in LLVM IR, globals can have an (optional) alignment attribute using keyword alignment. The integer value of the alignment must be a positive integer that is a power of 2.

Examples:

// Alignment is optional
llvm.mlir.global private constant @y(dense<1.0> : tensor<8xf32>) { alignment = 32 : i64 } : !llvm.array<8 x f32>

icmp

inline_asm

The InlineAsmOp mirrors the underlying LLVM semantics with a notable exception: the embedded asm_string is not allowed to define or reference any symbol or any global variable: only the operands of the op may be read, written, or referenced. Attempting to define or reference any symbol or any global behavior is considered undefined behavior at this time.

inline_asm :: MonadBlockBuilder m => Maybe Type -> [Value] -> ByteString -> ByteString -> m Value

A builder for llvm.inline_asm.

insertelement

pattern InsertElement :: Location -> Type -> operand -> operand -> operand -> AbstractOperation operand

A pattern for llvm.insertelement.

insertelement :: MonadBlockBuilder m => Type -> Value -> Value -> Value -> m Value

A builder for llvm.insertelement.

insertvalue

pattern InsertValue :: Location -> Type -> operand -> operand -> [Attribute] -> AbstractOperation operand

A pattern for llvm.insertvalue.

insertvalue :: MonadBlockBuilder m => Type -> Value -> Value -> [Attribute] -> m Value

A builder for llvm.insertvalue.

inttoptr

pattern IntToPtr :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.inttoptr.

inttoptr :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.inttoptr.

invoke

func

MLIR functions are defined by an operation that is not built into the IR itself. The LLVM dialect provides an llvm.func operation to define functions compatible with LLVM IR. These functions have LLVM dialect function type but use MLIR syntax to express it. They are required to have exactly one result type. LLVM function operation is intended to capture additional properties of LLVM functions, such as linkage and calling convention, that may be modeled differently by the built-in MLIR function.

// The type of @bar is !llvm<"i64 (i64)">
llvm.func @bar(%arg0: i64) -> i64 {
  llvm.return %arg0 : i64
}

// Type type of @foo is !llvm<"void (i64)">
// !llvm.void type is omitted
llvm.func @foo(%arg0: i64) {
  llvm.return
}

// A function with internal linkage.
llvm.func internal @internal_func() {
  llvm.return
}

func :: MonadBlockBuilder m => Maybe [Attribute] -> RegionBuilderT m () -> m ()

A builder for llvm.func.

lshr

pattern LShr :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.lshr.

lshr :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.lshr.

landingpad

pattern Landingpad :: Location -> Type -> [operand] -> AbstractOperation operand

A pattern for llvm.landingpad.

landingpad :: MonadBlockBuilder m => Type -> [Value] -> m Value

A builder for llvm.landingpad.

load

pattern Load :: Location -> Type -> operand -> Maybe Int -> AbstractOperation operand

A pattern for llvm.load.

load :: MonadBlockBuilder m => Type -> Value -> Maybe Int -> m Value

A builder for llvm.load.

intr.log10

pattern Log10 :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.log10.

intr_log10 :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.log10.

intr.log2

pattern Log2 :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.log2.

intr_log2 :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.log2.

intr.log

pattern Log :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.log.

intr_log :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.log.

intr.masked.load

intr_masked_load :: MonadBlockBuilder m => Type -> Value -> Value -> [Value] -> Int -> m Value

A builder for llvm.intr.masked.load.

intr.masked.store

pattern MaskedStore :: Location -> operand -> operand -> operand -> Int -> AbstractOperation operand

A pattern for llvm.intr.masked.store.

intr_masked_store :: MonadBlockBuilder m => Value -> Value -> Value -> Int -> m ()

A builder for llvm.intr.masked.store.

intr.matrix.column.major.load

intr.matrix.column.major.store

intr.matrix.multiply

pattern MatrixMultiply :: Location -> Type -> operand -> operand -> Int -> Int -> Int -> AbstractOperation operand

A pattern for llvm.intr.matrix.multiply.

intr_matrix_multiply :: MonadBlockBuilder m => Type -> Value -> Value -> Int -> Int -> Int -> m Value

A builder for llvm.intr.matrix.multiply.

intr.matrix.transpose

pattern MatrixTranspose :: Location -> Type -> operand -> Int -> Int -> AbstractOperation operand

A pattern for llvm.intr.matrix.transpose.

intr_matrix_transpose :: MonadBlockBuilder m => Type -> Value -> Int -> Int -> m Value

A builder for llvm.intr.matrix.transpose.

intr.maxnum

pattern MaxNum :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.maxnum.

intr_maxnum :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.intr.maxnum.

intr.maximum

pattern Maximum :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.maximum.

intr_maximum :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.intr.maximum.

intr.memcpy.inline

pattern MemcpyInline :: Location -> operand -> operand -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.memcpy.inline.

intr_memcpy_inline :: MonadBlockBuilder m => Value -> Value -> Value -> Value -> m ()

A builder for llvm.intr.memcpy.inline.

intr.memcpy

pattern Memcpy :: Location -> operand -> operand -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.memcpy.

intr_memcpy :: MonadBlockBuilder m => Value -> Value -> Value -> Value -> m ()

A builder for llvm.intr.memcpy.

intr.memset

pattern Memset :: Location -> operand -> operand -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.memset.

intr_memset :: MonadBlockBuilder m => Value -> Value -> Value -> Value -> m ()

A builder for llvm.intr.memset.

metadata

llvm.metadata op defines one or more metadata nodes.

Example: llvm.metadata @metadata { llvm.access_group @group1 llvm.access_group @group2 llvm.return }

intr.minnum

pattern MinNum :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.minnum.

intr_minnum :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.intr.minnum.

intr.minimum

pattern Minimum :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.minimum.

intr_minimum :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.intr.minimum.

mul

pattern Mul :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.mul.

mul :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.mul.

mlir.null

Unlike LLVM IR, MLIR does not have first-class null pointers. They must be explicitly created as SSA values using llvm.mlir.null. This operation has no operands or attributes, and returns a null value of a wrapped LLVM IR pointer type.

Examples:

// Null pointer to i8.
%0 = llvm.mlir.null : !llvm.ptr<i8>

// Null pointer to a function with signature void().
%1 = llvm.mlir.null : !llvm.ptr<func<void ()>>

pattern Null :: Location -> Type -> AbstractOperation operand

A pattern for llvm.mlir.null.

mlir_null :: MonadBlockBuilder m => Type -> m Value

A builder for llvm.mlir.null.

or

pattern Or :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.or.

or :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.or.

intr.pow

pattern Pow :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.pow.

intr_pow :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.intr.pow.

intr.prefetch

intr_prefetch :: MonadBlockBuilder m => Value -> Value -> Value -> Value -> m ()

A builder for llvm.intr.prefetch.

ptrtoint

pattern PtrToInt :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.ptrtoint.

ptrtoint :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.ptrtoint.

resume

pattern Resume :: Location -> operand -> AbstractOperation operand

A pattern for llvm.resume.

resume :: MonadBlockBuilder m => Value -> m EndOfBlock

A builder for llvm.resume.

return

pattern Return :: Location -> [operand] -> AbstractOperation operand

A pattern for llvm.return.

return :: MonadBlockBuilder m => [Value] -> m EndOfBlock

A builder for llvm.return.

intr.sadd.with.overflow

pattern SAddWithOverflow :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.sadd.with.overflow.

intr_sadd_with_overflow :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.intr.sadd.with.overflow.

sdiv

pattern SDiv :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.sdiv.

sdiv :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.sdiv.

sext

pattern SExt :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.sext.

sext :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.sext.

sitofp

pattern SIToFP :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.sitofp.

sitofp :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.sitofp.

intr.smax

pattern SMax :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.smax.

intr_smax :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.intr.smax.

intr.smin

pattern SMin :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.smin.

intr_smin :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.intr.smin.

intr.smul.with.overflow

pattern SMulWithOverflow :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.smul.with.overflow.

intr_smul_with_overflow :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.intr.smul.with.overflow.

srem

pattern SRem :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.srem.

srem :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.srem.

intr.ssub.with.overflow

pattern SSubWithOverflow :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.ssub.with.overflow.

intr_ssub_with_overflow :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.intr.ssub.with.overflow.

select

pattern Select :: Location -> Type -> operand -> operand -> operand -> AbstractOperation operand

A pattern for llvm.select.

select :: MonadBlockBuilder m => Type -> Value -> Value -> Value -> m Value

A builder for llvm.select.

shl

pattern Shl :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.shl.

shl :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.shl.

shufflevector

pattern ShuffleVector :: Location -> Type -> operand -> operand -> [Attribute] -> AbstractOperation operand

A pattern for llvm.shufflevector.

shufflevector :: MonadBlockBuilder m => Type -> Value -> Value -> [Attribute] -> m Value

A builder for llvm.shufflevector.

intr.sin

pattern Sin :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.sin.

intr_sin :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.sin.

intr.sqrt

pattern Sqrt :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.intr.sqrt.

intr_sqrt :: MonadBlockBuilder m => Value -> m Value

A builder for llvm.intr.sqrt.

intr.stackrestore

pattern StackRestore :: Location -> operand -> AbstractOperation operand

A pattern for llvm.intr.stackrestore.

intr_stackrestore :: MonadBlockBuilder m => Value -> m ()

A builder for llvm.intr.stackrestore.

intr.stacksave

pattern StackSave :: Location -> Type -> AbstractOperation operand

A pattern for llvm.intr.stacksave.

intr_stacksave :: MonadBlockBuilder m => Type -> m Value

A builder for llvm.intr.stacksave.

store

pattern Store :: Location -> operand -> operand -> Maybe Int -> AbstractOperation operand

A pattern for llvm.store.

store :: MonadBlockBuilder m => Value -> Value -> Maybe Int -> m ()

A builder for llvm.store.

sub

pattern Sub :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.sub.

sub :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.sub.

switch

trunc

pattern Trunc :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.trunc.

trunc :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.trunc.

intr.uadd.with.overflow

pattern UAddWithOverflow :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.uadd.with.overflow.

intr_uadd_with_overflow :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.intr.uadd.with.overflow.

udiv

pattern UDiv :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.udiv.

udiv :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.udiv.

uitofp

pattern UIToFP :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.uitofp.

uitofp :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.uitofp.

intr.umul.with.overflow

pattern UMulWithOverflow :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.umul.with.overflow.

intr_umul_with_overflow :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.intr.umul.with.overflow.

urem

pattern URem :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.urem.

urem :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.urem.

intr.usub.with.overflow

pattern USubWithOverflow :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.intr.usub.with.overflow.

intr_usub_with_overflow :: MonadBlockBuilder m => Type -> Value -> Value -> m Value

A builder for llvm.intr.usub.with.overflow.

mlir.undef

Unlike LLVM IR, MLIR does not have first-class undefined values. Such values must be created as SSA values using llvm.mlir.undef. This operation has no operands or attributes. It creates an undefined value of the specified LLVM IR dialect type wrapping an LLVM IR structure type.

Example:

// Create a structure with a 32-bit integer followed by a float.
%0 = llvm.mlir.undef : !llvm.struct<(i32, f32)>

pattern Undef :: Location -> Type -> AbstractOperation operand

A pattern for llvm.mlir.undef.

mlir_undef :: MonadBlockBuilder m => Type -> m Value

A builder for llvm.mlir.undef.

unreachable

pattern Unreachable :: Location -> AbstractOperation operand

A pattern for llvm.unreachable.

unreachable :: MonadBlockBuilder m => m EndOfBlock

A builder for llvm.unreachable.

xor

pattern XOr :: Location -> Type -> operand -> operand -> AbstractOperation operand

A pattern for llvm.xor.

xor :: MonadBlockBuilder m => Value -> Value -> m Value

A builder for llvm.xor.

zext

pattern ZExt :: Location -> Type -> operand -> AbstractOperation operand

A pattern for llvm.zext.

zext :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.zext.

intr.masked.compressstore

intr_masked_compressstore :: MonadBlockBuilder m => Value -> Value -> Value -> m ()

A builder for llvm.intr.masked.compressstore.

intr.masked.expandload

intr_masked_expandload :: MonadBlockBuilder m => Type -> Value -> Value -> Value -> m Value

A builder for llvm.intr.masked.expandload.

intr.masked.gather

intr_masked_gather :: MonadBlockBuilder m => Type -> Value -> Value -> [Value] -> Int -> m Value

A builder for llvm.intr.masked.gather.

intr.masked.scatter

intr_masked_scatter :: MonadBlockBuilder m => Value -> Value -> Value -> Int -> m ()

A builder for llvm.intr.masked.scatter.

intr.vector.reduce.add

intr_vector_reduce_add :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.vector.reduce.add.

intr.vector.reduce.and

intr_vector_reduce_and :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.vector.reduce.and.

intr.vector.reduce.fadd

intr_vector_reduce_fadd :: MonadBlockBuilder m => Type -> Value -> Value -> Bool -> m Value

A builder for llvm.intr.vector.reduce.fadd.

intr.vector.reduce.fmax

intr_vector_reduce_fmax :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.vector.reduce.fmax.

intr.vector.reduce.fmin

intr_vector_reduce_fmin :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.vector.reduce.fmin.

intr.vector.reduce.fmul

intr_vector_reduce_fmul :: MonadBlockBuilder m => Type -> Value -> Value -> Bool -> m Value

A builder for llvm.intr.vector.reduce.fmul.

intr.vector.reduce.mul

intr_vector_reduce_mul :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.vector.reduce.mul.

intr.vector.reduce.or

intr_vector_reduce_or :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.vector.reduce.or.

intr.vector.reduce.smax

intr_vector_reduce_smax :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.vector.reduce.smax.

intr.vector.reduce.smin

intr_vector_reduce_smin :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.vector.reduce.smin.

intr.vector.reduce.umax

intr_vector_reduce_umax :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.vector.reduce.umax.

intr.vector.reduce.umin

intr_vector_reduce_umin :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.vector.reduce.umin.

intr.vector.reduce.xor

intr_vector_reduce_xor :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for llvm.intr.vector.reduce.xor.