MLIR.AST.Dialect.Linalg

index

The linalg.index operation returns the iteration index of the immediately enclosing linalg structured operation for the iteration dimension dim. The dim attribute specifies the position of the accessed dimension in the indexing map domain.

Example:

#map = affine_map<(i, j) -> (i, j)>
linalg.generic {indexing_maps = [#map, #map],
                iterator_types = ["parallel", "parallel"]}
  outs(%I, %J : memref<?x?xindex>, memref<?x?xindex>) {
  ^bb0(%arg0 : index, %arg1 : index):
  // Access the outer iteration dimension i
  %i = linalg.index 0 : index
  // Access the inner iteration dimension j
  %j = linalg.index 1 : index
  linalg.yield %i, %j : index, index
}

This may lower to IR resembling:

%0 = dim %I, %c0 : memref<?x?xindex>
%1 = dim %I, %c1 : memref<?x?xindex>
scf.for %i = %c0 to %0 step %c1 {
  scf.for %j = %c0 to %1 step %c1 {
    store %i, %I[%i, %j] : memref<?x?xindex>
    store %j, %J[%i, %j] : memref<?x?xindex>
  }
}

pattern Linalg_Index :: Location -> Type -> Int -> AbstractOperation operand

A pattern for linalg.index.

index :: MonadBlockBuilder m => Type -> Int -> m Value

A builder for linalg.index.

init_tensor

linalg.init_tensor is an operation that materializes a tensor of a given shape. The shape could be dynamic or static.

pattern Linalg_InitTensor :: Location -> Type -> [operand] -> [Int] -> AbstractOperation operand

A pattern for linalg.init_tensor.

init_tensor :: MonadBlockBuilder m => Type -> [Value] -> [Int] -> m Value

A builder for linalg.init_tensor.

yield

linalg.yield is a special terminator operation for blocks inside regions in linalg generic ops. It returns values to the immediately enclosing linalg generic op.

Example:

linalg.yield %f0, %f1 : f32, f32

pattern Linalg_Yield :: Location -> [operand] -> AbstractOperation operand

A pattern for linalg.yield.

yield :: MonadBlockBuilder m => [Value] -> m EndOfBlock

A builder for linalg.yield.

batch_matmul

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

batch_matmul :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.batch_matmul.

batch_matvec

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

batch_matvec :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.batch_matvec.

conv_1d_nwc_wcf

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_1d_nwc_wcf :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.conv_1d_nwc_wcf.

conv_1d

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_1d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.conv_1d.

conv_2d_nchw_fchw

Layout: * Input: NCHW. * Kernel: FCHW.

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_2d_nchw_fchw :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.conv_2d_nchw_fchw.

conv_2d_nhwc_hwcf

Layout: * Input: NHWC. * Kernel: HWCF.

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_2d_nhwc_hwcf :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.conv_2d_nhwc_hwcf.

conv_2d_nhwc_hwcf_q

Layout: * Input: NHWC. * Kernel: HWCF.

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output. This includes the zero point offsets common to quantized operations.

conv_2d_nhwc_hwcf_q :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.conv_2d_nhwc_hwcf_q.

conv_2d

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_2d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.conv_2d.

conv_3d_ndhwc_dhwcf

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_3d_ndhwc_dhwcf :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.conv_3d_ndhwc_dhwcf.

conv_3d

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_3d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.conv_3d.

depthwise_conv_1d_nwc_wc

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output. Multiplier is set to 1 which is a special case for most depthwise convolutions.

depthwise_conv_1d_nwc_wc :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.depthwise_conv_1d_nwc_wc.

depthwise_conv_2d_nhwc_hwc

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output. Multiplier is set to 1 which is a special case for most depthwise convolutions.

depthwise_conv_2d_nhwc_hwc :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.depthwise_conv_2d_nhwc_hwc.

depthwise_conv_2d_nhwc_hwc_q

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

depthwise_conv_2d_nhwc_hwc_q :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.depthwise_conv_2d_nhwc_hwc_q.

depthwise_conv_2d_nhwc_hwcm

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

depthwise_conv_2d_nhwc_hwcm :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.depthwise_conv_2d_nhwc_hwcm.

depthwise_conv_2d_nhwc_hwcm_q

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

depthwise_conv_2d_nhwc_hwcm_q :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.depthwise_conv_2d_nhwc_hwcm_q.

dot

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

dot :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.dot.

elemwise_binary

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

elemwise_binary :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.elemwise_binary.

elemwise_unary

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

elemwise_unary :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.elemwise_unary.

fill

fill :: MonadBlockBuilder m => Maybe Type -> Value -> Value -> RegionBuilderT m () -> m Value

A builder for linalg.fill.

fill_rng_2d

The operation generations pseudo random numbers using a linear congruential generator. It provides no guarantees regarding the distribution of the generated random numbers. Instead of generating the random numbers sequentially, it instantiates one random number generator per data element and runs them in parallel. The seed operand and the indices of the data element seed the random number generation. The min and max operands limit the range of the generated random numbers.

fill_rng_2d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.fill_rng_2d.

fill_tensor

Works for arbitrary ranked output tensors since the operation performs scalar accesses only and is thus rank polymorphic. Numeric casting is performed on the value operand, promoting it to the same data type as the output.

fill_tensor :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.fill_tensor.

generic

Generic Linalg op form where the key properties of the computation are specified as attributes. In pretty form, a linalg.generic op is written as:

  linalg.generic #trait_attribute
      ins(%A, %B : memref<?x?xf32, stride_specification>,
                   memref<?x?xf32, stride_specification>)
      outs(%C : memref<?x?xf32, stride_specification>)
      attrs = {other-optional-attributes}
      {region}
  

Where #trait_attributes is an alias of a dictionary attribute containing: - doc [optional]: a documentation string - indexing_maps: a list of AffineMapAttr, one AffineMapAttr per each input and output view. Such AffineMapAttr specifies the mapping between the loops and the indexing within each view. - library_call [optional]: a StringAttr containing the name of an external library function that the linalg.generic operation maps to. The external library is assumed to be dynamically linked and no strong compile-time guarantees are provided. In the absence of such a library call, linalg.generic will always lower to loops. - iterator_types: an ArrayAttr specifying the type of the enclosing loops. Each element of the list represents and iterator of one of the following types: parallel, reduction, window

Example: Defining a #matmul_trait attribute in MLIR can be done as follows: #matmul_accesses = [ (m, n, k) -> (m, k), (m, n, k) -> (k, n), (m, n, k) -> (m, n) ] #matmul_trait = { doc = "C(m, n) += A(m, k) * B(k, n)", indexing_maps = #matmul_accesses, library_call = "linalg_matmul", iterator_types = ["parallel", "parallel", "reduction"] }

And can be reused in multiple places as: linalg.generic #matmul_trait ins(%A, %B : memref<?x?xf32, stride_specification>, memref<?x?xf32, stride_specification>) outs(%C : memref<?x?xf32, stride_specification>) {other-optional-attributes} { ^bb0(%a: f32, %b: f32, %c: f32) : %d = arith.mulf %a, %b: f32 %e = arith.addf %c, %d: f32 linalg.yield %e : f32 }

This may lower to either: call @linalg_matmul(%A, %B, %C) : (memref<?x?xf32, stride_specification>, memref<?x?xf32, stride_specification>, memref<?x?xf32, stride_specification>) -> ()

or IR resembling: scf.for %m = %c0 to %M step %c1 { scf.for %n = %c0 to %N step %c1 { scf.for %k = %c0 to %K step %c1 { %a = load %A[%m, %k] : memref<?x?xf32, stride_specification> %b = load %B[%k, %n] : memref<?x?xf32, stride_specification> %c = load %C[%m, %n] : memref<?x?xf32, stride_specification> %d = arith.mulf %a, %b: f32 %e = arith.addf %c, %d: f32 store %e, %C[%m, %n] : memref<?x?x?xf32, stride_specification> } } }

To allow progressive lowering from the value world (a.k.a tensor values) to the buffer world (a.k.a memref values), a linalg.generic op allows mixing tensors and buffers operands and tensor results.

%C = linalg.generic #trait_attribute
  ins(%A, %B : tensor<?x?xf32>, memref<?x?xf32, stride_specification>)
  outs(%C : tensor<?x?xf32>)
  {other-optional-attributes}
  {region}
  -> (tensor<?x?xf32>)

generic :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> [Map] -> [Attribute] -> Maybe ByteString -> Maybe ByteString -> RegionBuilderT m () -> m Value

A builder for linalg.generic.

matmul

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

matmul :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.matmul.

matmul_unsigned

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

matmul_unsigned :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.matmul_unsigned.

matvec

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

matvec :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.matvec.

mmt4d

Differences from linalg.matmul: * The right hand side is transposed, whence the 't' in 'mmt'. * The input and output tensors have a 4D shape instead of a 2D shape. They are interpreted as 2D matrices with one level of 2D tile subdivision, whence the 2+2=4 dimensions. The inner tile dimensions are identified with '0' suffixes below, for instance the LHS matrix shape (M, K, M0, K0) reads as: MxK tiles, each of shape M0xK0.

mmt4d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.mmt4d.

pooling_nchw_max

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_nchw_max :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.pooling_nchw_max.

pooling_ndhwc_max

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_ndhwc_max :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.pooling_ndhwc_max.

pooling_ndhwc_min

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_ndhwc_min :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.pooling_ndhwc_min.

pooling_ndhwc_sum

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_ndhwc_sum :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.pooling_ndhwc_sum.

pooling_nhwc_max

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_nhwc_max :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.pooling_nhwc_max.

pooling_nhwc_max_unsigned

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_nhwc_max_unsigned :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.pooling_nhwc_max_unsigned.

pooling_nhwc_min

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_nhwc_min :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.pooling_nhwc_min.

pooling_nhwc_min_unsigned

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_nhwc_min_unsigned :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.pooling_nhwc_min_unsigned.

pooling_nhwc_sum

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_nhwc_sum :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.pooling_nhwc_sum.

quantized_batch_matmul

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output. The quantized variant includes zero-point adjustments for the left and right operands of the matmul.

quantized_batch_matmul :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.quantized_batch_matmul.

quantized_matmul

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output. The quantized variant includes zero-point adjustments for the left and right operands of the matmul.

quantized_matmul :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.quantized_matmul.

soft_plus_2d

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

soft_plus_2d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.soft_plus_2d.

vecmat

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

vecmat :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.vecmat.