MLIR.AST.Dialect.Linalg

index

The linalg.index operation returns the iteration index of the immediately enclosing linalg structured operation for the iteration dimension dim. The dim attribute specifies the position of the accessed dimension in the indexing map domain.

Example:

#map = affine_map<(i, j) -> (i, j)>
linalg.generic {indexing_maps = [#map, #map],
                iterator_types = ["parallel", "parallel"]}
  outs(%I, %J : memref<?x?xindex>, memref<?x?xindex>) {
  ^bb0(%arg0 : index, %arg1 : index):
  // Access the outer iteration dimension i
  %i = linalg.index 0 : index
  // Access the inner iteration dimension j
  %j = linalg.index 1 : index
  linalg.yield %i, %j : index, index
}

This may lower to IR resembling:

%0 = dim %I, %c0 : memref<?x?xindex>
%1 = dim %I, %c1 : memref<?x?xindex>
scf.for %i = %c0 to %0 step %c1 {
  scf.for %j = %c0 to %1 step %c1 {
    store %i, %I[%i, %j] : memref<?x?xindex>
    store %j, %J[%i, %j] : memref<?x?xindex>
  }
}

pattern Linalg_Index :: Location -> Type -> Int -> AbstractOperation operand

A pattern for linalg.index.

index :: MonadBlockBuilder m => Type -> Int -> m Value

A builder for linalg.index.

init_tensor

linalg.init_tensor is an operation that materializes a tensor of a given shape. The shape could be dynamic or static.

pattern Linalg_InitTensor :: Location -> Type -> [operand] -> [Int] -> AbstractOperation operand

A pattern for linalg.init_tensor.

init_tensor :: MonadBlockBuilder m => Type -> [Value] -> [Int] -> m Value

A builder for linalg.init_tensor.

pad_tensor

linalg.pad_tensor is an operation that pads the source tensor with given low and high padding config.

The PadTensor operation supports the following arguments:

source: the "base" tensor on which to pad.

low: A list contains the padding along the start of each dimension, i.e low.

high: A list contains the padding along the end of each dimension, i.e. high.

nofold: indicates that the operation should not be folded when source and result types are equal.

The result tensor dimensions are low + dim + high along that dimension. The number of elements of low and high must match the rank of the input tensor. They can be either a constant or a dynamic value.

The region of the pad_tensor operation returns the value to use for the padding. The arguments of the region represent the index of the source being accessed. There should be as many arguments as the rank of the source tensor. The value yield-ed by the region is used as the value of the view at the given position.

If nofold is set, the padding operation will not be folded away even if the source type and the padded type have the same static shape. This can be used, e.g., for packing or promotion to faster memory.

Example 1:

  %pad_value = ... : f32
  %0 = linalg.pad_tensor %0 low[1, 2] high[2, 3] {
  ^bb0(%arg0 : index, %arg1 : index):
    linalg.yield %pad_value : f32
  } : tensor<?x?xf32> to tensor<?x?xf32>

Example 2:

  %pad_value = ... : f32
  %0 = linalg.pad_tensor %arg0 low[2, %arg1, 3, 3] high[3, 3, %arg1, 2] {
  ^bb0(%arg2: index, %arg3: index, %arg4: index, %arg5: index):
      linalg.yield %pad_value : f32
  } : tensor<1x2x2x?xf32> to tensor<6x?x?x?xf32>

Example 3:

  %pad_value = ... : f32
  %0 = linalg.pad_tensor %arg0 low[0, 0] high[%ub0, %ub1] {
  ^bb0(%arg1: index, %arg2: index):
    linalg.yield %pad_value : f32
  } : tensor<2x3xf32> to tensor<?x?xf32>

Example 4:

  // Force a padded value to be always exist with nofold.
  %pad_value = ... : f32
  %0 = linalg.pad_tensor %arg0 nofold low[0, 0] high[0, 0] {
  ^bb0(%arg1: index, %arg2: index):
    linalg.yield %pad_value : f32
  } : tensor<2x3xf32> to tensor<2x3xf32>

pad_tensor :: MonadBlockBuilder m => Type -> Value -> [Value] -> [Value] -> [Int] -> [Int] -> RegionBuilderT m () -> m Value

A builder for linalg.pad_tensor.

range

The linalg.range op creates a !linalg.range from 3 values of type index that represent the min, max and step values of the range. This type does not pass function boundaries at the moment.

Example:

%3 = linalg.range %0:%1:%2 : !linalg.range

@

pattern Linalg_Range :: Location -> Type -> operand -> operand -> operand -> AbstractOperation operand

A pattern for linalg.range.

range :: MonadBlockBuilder m => Type -> Value -> Value -> Value -> m Value

A builder for linalg.range.

tensor_collapse_shape

The linalg.tensor_collapse_shape op produces a new tensor with a smaller rank whose sizes are a reassociation of the original src.

A reassociation is defined as a continuous grouping of dimensions and is represented with an array of I64ArrayAttr attribute.

The verification rule is that the reassociation maps are applied to the operand tensor with the higher rank to obtain the result tensor with the smaller rank.

The result tensor type of a reshape can be zero-ranked if the operand tensor type is statically shaped with all dimensions being unit extent. In such case the reassociation map is empty.

Examples:

// Dimension collapse (i, j) -> i' and k -> k'
%b = linalg.tensor_collapse_shape %a [[0, 1], [2]]
    : tensor<?x?x?xf32> into tensor<?x?xf32>

tensor_expand_shape

The linalg.tensor_expand_shape op produces a new tensor with a higher rank whose sizes are a reassociation of the original src.

A reassociation is defined as a continuous grouping of dimensions and is represented with an array of I64ArrayAttr attribute.

The verification rule is that the reassociation maps are applied to the result tensor with the higher rank to obtain the operand tensor with the smaller rank.

The operand tensor type of a reshape can be zero-ranked if the result tensor type is statically shaped with all dimensions being unit extent. In such cases the reassociation map is empty.

Examples:

// Dimension expansion i -> (i', j') and (k) -> (k')
%b = linalg.tensor_expand_shape %a [[0, 1], [2]]
    : tensor<?x?xf32> into tensor<?x?x?xf32>

tiled_loop

This is a loop-like operation with additional properties. The arguments also include the input and the output tensors or memrefs and the attributes to specify the iterator types.

Parsing TiledLoopOp will set all elements of the iterator_types attribute to "parallel" type, when it is absent from the custom format.

Tensor-based version:

The body region of the loop contains extract_slice operations applied to every tensor argument of TiledLoopOp.

The body region must contain exactly one block that terminates with linalg.yield with the operands resulting from insert_slice operations.

Example:

%0 = linalg.tiled_loop (%i) = (%c0) to (%c24) step (%c4)
    ins(%lhs, %rhs : tensor<24x64xi8>, tensor<24x64xi8>)
    outs(%out : tensor<24x64xi8>)
    iterators("parallel")
    distribution("block_x") {
  %lhs_sub = tensor.extract_slice %lhs[%i, 0] [%c4, %c64] [1, 1]
      : tensor<24x64xi8> to tensor<?x?xi8>
  %rhs_sub = tensor.extract_slice %rhs[%i, 0] [%c4, %c64] [1, 1]
      : tensor<24x64xi8> to tensor<?x?xi8>
  %out_sub = tensor.extract_slice %out[%i, 0] [%c4, %c64] [1, 1]
      : tensor<24x64xi8> to tensor<?x?xi8>

  %result_sub = linalg.generic ...

  %result = tensor.insert_slice %result_sub into %out[%i, 0][%c4, %c64][1, 1]
    : tensor<?x?xi8> into tensor<24x64xi8>
  linalg.yield %result : tensor<24x64xi8>
}

MemRef-based version:

The body region of the loop contains subview operations applied to every memref argument of TiledLoopOp.

The body region must contain exactly one block that terminates with linalg.yield with no operands.

Example:

linalg.tiled_loop (%i) = (%c0) to (%c24) step (%c4)
    ins(%lhs, %rhs : memref<24x64xi8>, memref<24x64xi8>)
    outs(%out : memref<24x64xi8>)
    iterators("parallel")
    distribution("block_x") {
  %lhs_sub = subview %lhs[%i, 0] [%c4, %c64] [1, 1]
      : memref<24x64xi8> to memref<?x?xi8>
  %rhs_sub = subview %rhs[%i, 0] [%c4, %c64] [1, 1]
      : memref<24x64xi8> to memref<?x?xi8>
  %out_sub = subview %out[%i, 0] [%c4, %c64] [1, 1]
      : memref<24x64xi8> to memref<?x?xi8>

  %result_sub = linalg.generic ...
  linalg.yield
}

tiled_loop :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> [Value] -> [Value] -> [Value] -> [Attribute] -> Maybe [Attribute] -> RegionBuilderT m () -> m Value

A builder for linalg.tiled_loop.

yield

linalg.yield is a special terminator operation for blocks inside regions in linalg generic ops. It returns values to the immediately enclosing linalg generic op.

Example:

linalg.yield %f0, %f1 : f32, f32

pattern Linalg_Yield :: Location -> [operand] -> AbstractOperation operand

A pattern for linalg.yield.

yield :: MonadBlockBuilder m => [Value] -> m EndOfBlock

A builder for linalg.yield.

batch_matmul

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

batch_matmul :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.batch_matmul.

batch_matvec

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

batch_matvec :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.batch_matvec.

conv_1d_nwc_wcf

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_1d

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_1d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.conv_1d.

conv_2d_nchw_fchw

Layout: * Input: NCHW. * Kernel: FCHW.

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_2d_nhwc_hwcf

Layout: * Input: NHWC. * Kernel: HWCF.

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_2d_nhwc_hwcf_q

Layout: * Input: NHWC. * Kernel: HWCF.

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output. This includes the zero point offsets common to quantized operations.

conv_2d

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_2d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.conv_2d.

conv_3d_ndhwc_dhwcf

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_3d

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

conv_3d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.conv_3d.

copy

Copies the data in the input view into the output view.

Usage:

linalg.copy(%arg0, %arg1) : memref<?xf32, stride_specification>,
                            memref<?xf32, stride_specification>

One possible lowering to loop form is:

%0 = linalg.dim %arg0, 0 : index
scf.for %i0 = %c0 to %0 step %c1 {
  %1 = load %arg0[%i0] : memref<?xf32, stride_specification>
  store %1, %arg1[%i0] : memref<?xf32, stride_specification>
}

Optionally, can take input_permutation and output_permutation attributes to reorder the dimensions of the input and output views.

Usage:

linalg.copy(%arg0, %arg1) {inputPermutation : (i, j, k) -> (i, k, j),
                           outputPermutation : (i, j, k) -> (k, j, i)} :
  memref<?x?x?xf32, stride_specification>,
  memref<?x?x?xf32, stride_specification>

One possible lowering to loop form is:

%0 = linalg.dim %arg0, 0
%1 = linalg.dim %arg0, 1
%2 = linalg.dim %arg0, 2
scf.for %i0 = %c0 to %{{.*}} step %c1 {
  scf.for %i1 = %c0 to %{{.*}} step %c1 {
    scf.for %i2 = %c0 to %{{.*}} step %c1 {
      %3 = load %arg0[%i0, %i2, %i1] :
              memref<?x?x?xf32, stride_specification>
      store %3, %arg1[%i2, %i1, %i0] :
              memref<?x?x?xf32, stride_specification>

The views are expected to be compatible for correctness but this is not enforced at the moment.

copy :: MonadBlockBuilder m => Value -> Value -> Maybe Map -> Maybe Map -> RegionBuilderT m () -> m ()

A builder for linalg.copy.

depthwise_conv_1d_nwc_wc

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output. Multiplier is set to 1 which is a special case for most depthwise convolutions.

depthwise_conv_2d_nhwc_hwc

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output. Multiplier is set to 1 which is a special case for most depthwise convolutions.

depthwise_conv_2d_nhwc_hwc_q

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

depthwise_conv_2d_nhwc_hwcm

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

depthwise_conv_2d_nhwc_hwcm_q

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

dot

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

dot :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.dot.

fill

fill :: MonadBlockBuilder m => Maybe Type -> Value -> Value -> RegionBuilderT m () -> m Value

A builder for linalg.fill.

fill_rng_2d

The operation generations pseudo random numbers using a linear congruential generator. It provides no guarantees regarding the distribution of the generated random numbers. Instead of generating the random numbers sequentially, it instantiates one random number generator per data element and runs them in parallel. The seed operand and the indices of the data element seed the random number generation. The min and max operands limit the range of the generated random numbers.

fill_rng_2d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.fill_rng_2d.

generic

Generic Linalg op form where the key properties of the computation are specified as attributes. In pretty form, a linalg.generic op is written as:

  linalg.generic #trait_attribute
      ins(%A, %B : memref<?x?xf32, stride_specification>,
                   memref<?x?xf32, stride_specification>)
      outs(%C : memref<?x?xf32, stride_specification>)
      attrs = {other-optional-attributes}
      {region}
  

Where #trait_attributes is an alias of a dictionary attribute containing: - doc [optional]: a documentation string - indexing_maps: a list of AffineMapAttr, one AffineMapAttr per each input and output view. Such AffineMapAttr specifies the mapping between the loops and the indexing within each view. - library_call [optional]: a StringAttr containing the name of an external library function that the linalg.generic operation maps to. The external library is assumed to be dynamically linked and no strong compile-time guarantees are provided. In the absence of such a library call, linalg.generic will always lower to loops. - iterator_types: an ArrayAttr specifying the type of the enclosing loops. Each element of the list represents and iterator of one of the following types: parallel, reduction, window

Example: Defining a #matmul_trait attribute in MLIR can be done as follows: #matmul_accesses = [ (m, n, k) -> (m, k), (m, n, k) -> (k, n), (m, n, k) -> (m, n) ] #matmul_trait = { doc = "C(m, n) += A(m, k) * B(k, n)", indexing_maps = #matmul_accesses, library_call = "linalg_matmul", iterator_types = ["parallel", "parallel", "reduction"] }

And can be reused in multiple places as: linalg.generic #matmul_trait ins(%A, %B : memref<?x?xf32, stride_specification>, memref<?x?xf32, stride_specification>) outs(%C : memref<?x?xf32, stride_specification>) {other-optional-attributes} { ^bb0(%a: f32, %b: f32, %c: f32) : %d = arith.mulf %a, %b: f32 %e = arith.addf %c, %d: f32 linalg.yield %e : f32 }

This may lower to either: call @linalg_matmul(%A, %B, %C) : (memref<?x?xf32, stride_specification>, memref<?x?xf32, stride_specification>, memref<?x?xf32, stride_specification>) -> ()

or IR resembling: scf.for %m = %c0 to %M step %c1 { scf.for %n = %c0 to %N step %c1 { scf.for %k = %c0 to %K step %c1 { %a = load %A[%m, %k] : memref<?x?xf32, stride_specification> %b = load %B[%k, %n] : memref<?x?xf32, stride_specification> %c = load %C[%m, %n] : memref<?x?xf32, stride_specification> %d = arith.mulf %a, %b: f32 %e = arith.addf %c, %d: f32 store %e, %C[%m, %n] : memref<?x?x?xf32, stride_specification> } } }

To allow progressive lowering from the value world (a.k.a tensor values) to the buffer world (a.k.a memref values), a linalg.generic op allows mixing tensors and buffers operands and tensor results.

%C = linalg.generic #trait_attribute
  ins(%A, %B : tensor<?x?xf32>, memref<?x?xf32, stride_specification>)
  outs(%C : tensor<?x?xf32>)
  {other-optional-attributes}
  {region}
  -> (tensor<?x?xf32>)

generic :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> [Map] -> [Attribute] -> Maybe ByteString -> Maybe ByteString -> RegionBuilderT m () -> m Value

A builder for linalg.generic.

matmul

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

matmul :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.matmul.

matmul_unsigned

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

matmul_unsigned :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.matmul_unsigned.

matvec

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

matvec :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.matvec.

mmt4d

Differences from linalg.matmul: * The right hand side is transposed, whence the 't' in 'mmt'. * The input and output tensors have a 4D shape instead of a 2D shape. They are interpreted as 2D matrices with one level of 2D tile subdivision, whence the 2+2=4 dimensions. The inner tile dimensions are identified with '0' suffixes below, for instance the LHS matrix shape (M, K, M0, K0) reads as: MxK tiles, each of shape M0xK0.

mmt4d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.mmt4d.

pooling_nchw_max

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_ndhwc_max

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_ndhwc_min

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_ndhwc_sum

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_nhwc_max

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_nhwc_max_unsigned

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_nhwc_min

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_nhwc_min_unsigned

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

pooling_nhwc_sum

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

quantized_batch_matmul

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output. The quantized variant includes zero-point adjustments for the left and right operands of the matmul.

quantized_batch_matmul :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.quantized_batch_matmul.

quantized_matmul

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output. The quantized variant includes zero-point adjustments for the left and right operands of the matmul.

quantized_matmul :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.quantized_matmul.

soft_plus_2d

Numeric casting is performed on the input operand, promoting it to the same data type as the accumulator/output.

soft_plus_2d :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.soft_plus_2d.

vecmat

Numeric casting is performed on the operands to the inner multiply, promoting them to the same data type as the accumulator/output.

vecmat :: MonadBlockBuilder m => [Type] -> [Value] -> [Value] -> RegionBuilderT m () -> m Value

A builder for linalg.vecmat.