MLIR.AST.Dialect.Std

Documentation

pattern Branch :: Location -> BlockName -> [Name] -> Operation

br :: MonadBlockBuilder m => BlockName -> [Value] -> m EndOfBlock

cond_br :: MonadBlockBuilder m => Value -> BlockName -> [Value] -> BlockName -> [Value] -> m EndOfBlock

assert

Assert operation with single boolean operand and an error message attribute. If the argument is true this operation has no effect. Otherwise, the program execution will abort. The provided error message may be used by a runtime to propagate the error to the user.

Example:

assert %b, "Expected ... to be true"

pattern Assert :: Location -> operand -> ByteString -> AbstractOperation operand

A pattern for std.assert.

assert :: MonadBlockBuilder m => Value -> ByteString -> m ()

A builder for std.assert.

atomic_rmw

The atomic_rmw operation provides a way to perform a read-modify-write sequence that is free from data races. The kind enumeration specifies the modification to perform. The value operand represents the new value to be applied during the modification. The memref operand represents the buffer that the read and write will be performed against, as accessed by the specified indices. The arity of the indices is the rank of the memref. The result represents the latest value that was stored.

Example:

%x = atomic_rmw "addf" %value, %I[%i] : (f32, memref<10xf32>) -> f32

atomic_yield

"atomic_yield" yields an SSA value from a GenericAtomicRMWOp region.

pattern AtomicYield :: Location -> operand -> AbstractOperation operand

A pattern for std.atomic_yield.

atomic_yield :: MonadBlockBuilder m => Value -> m EndOfBlock

A builder for std.atomic_yield.

br

The br operation represents a branch operation in a function. The operation takes variable number of operands and produces no results. The operand number and types for each successor must match the arguments of the block successor.

Example:

^bb2:
  %2 = call @someFn()
  br ^bb3(%2 : tensor<*xf32>)
^bb3(%3: tensor<*xf32>):

call_indirect

The call_indirect operation represents an indirect call to a value of function type. Functions are first class types in MLIR, and may be passed as arguments and merged together with block arguments. The operands and result types of the call must match the specified function type.

Function values can be created with the constant operation.

Example:

%31 = call_indirect %15(%0, %1)
        : (tensor<16xf32>, tensor<16xf32>) -> tensor<16xf32>

call_indirect :: MonadBlockBuilder m => [Type] -> Value -> [Value] -> m Value

A builder for std.call_indirect.

call

The call operation represents a direct call to a function that is within the same symbol scope as the call. The operands and result types of the call must match the specified function type. The callee is encoded as a symbol reference attribute named "callee".

Example:

%2 = call @my_add(%0, %1) : (f32, f32) -> f32

cond_br

The cond_br terminator operation represents a conditional branch on a boolean (1-bit integer) value. If the bit is set, then the first destination is jumped to; if it is false, the second destination is chosen. The count and types of operands must align with the arguments in the corresponding target blocks.

The MLIR conditional branch operation is not allowed to target the entry block for a region. The two destinations of the conditional branch operation are allowed to be the same.

The following example illustrates a function with a conditional branch operation that targets the same block.

Example:

func @select(%a: i32, %b: i32, %flag: i1) -> i32 {
  // Both targets are the same, operands differ
  cond_br %flag, ^bb1(%a : i32), ^bb1(%b : i32)

^bb1(%x : i32) :
  return %x : i32
}

constant

Syntax:

operation ::= ssa-id = std.constant attribute-value : type

The constant operation produces an SSA value equal to some constant specified by an attribute. This is the way that MLIR uses to form simple integer and floating point constants, as well as more exotic things like references to functions and tensor/vector constants.

Example:

// Complex constant
%1 = constant [1.0 : f32, 1.0 : f32] : complex<f32>

// Reference to function @myfn.
%2 = constant @myfn : (tensor<16xf32>, f32) -> tensor<16xf32>

// Equivalent generic forms
%1 = "std.constant"() {value = [1.0 : f32, 1.0 : f32] : complex<f32>}
   : () -> complex<f32>
%2 = "std.constant"() {value = @myfn}
   : () -> ((tensor<16xf32>, f32) -> tensor<16xf32>)

MLIR does not allow direct references to functions in SSA operands because the compiler is multithreaded, and disallowing SSA values to directly reference a function simplifies this (rationale.

pattern Constant :: Location -> Type -> Attribute -> AbstractOperation operand

A pattern for std.constant.

constant :: MonadBlockBuilder m => Type -> Attribute -> m Value

A builder for std.constant.

generic_atomic_rmw

The generic_atomic_rmw operation provides a way to perform a read-modify-write sequence that is free from data races. The memref operand represents the buffer that the read and write will be performed against, as accessed by the specified indices. The arity of the indices is the rank of the memref. The result represents the latest value that was stored. The region contains the code for the modification itself. The entry block has a single argument that represents the value stored in memref[indices] before the write is performed. No side-effecting ops are allowed in the body of GenericAtomicRMWOp.

Example:

%x = generic_atomic_rmw %I[%i] : memref<10xf32> {
  ^bb0(%current_value : f32):
    %c1 = arith.constant 1.0 : f32
    %inc = arith.addf %c1, %current_value : f32
    atomic_yield %inc : f32
}

generic_atomic_rmw :: MonadBlockBuilder m => Type -> Value -> [Value] -> RegionBuilderT m () -> m Value

A builder for std.generic_atomic_rmw.

rank

The rank operation takes a memref/tensor operand and returns its rank.

Example:

%1 = rank %arg0 : tensor<*xf32>
%2 = rank %arg1 : memref<*xf32>

pattern Rank :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.rank.

rank :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.rank.

return

The return operation represents a return operation within a function. The operation takes variable number of operands and produces no results. The operand number and types must match the signature of the function that contains the operation.

Example:

func @foo() : (i32, f8) {
  ...
  return %0, %1 : i32, f8
}

pattern Return :: Location -> [operand] -> AbstractOperation operand

A pattern for std.return.

return :: MonadBlockBuilder m => [Value] -> m EndOfBlock

A builder for std.return.

select

The select operation chooses one value based on a binary condition supplied as its first operand. If the value of the first operand is 1, the second operand is chosen, otherwise the third operand is chosen. The second and the third operand must have the same type.

The operation applies to vectors and tensors elementwise given the _shape_ of all operands is identical. The choice is made for each element individually based on the value at the same position as the element in the condition operand. If an i1 is provided as the condition, the entire vector or tensor is chosen.

The select operation combined with cmpi can be used to implement min and max with signed or unsigned comparison semantics.

Example:

// Custom form of scalar selection.
%x = select %cond, %true, %false : i32

// Generic form of the same operation.
%x = "std.select"(%cond, %true, %false) : (i1, i32, i32) -> i32

// Element-wise vector selection.
%vx = std.select %vcond, %vtrue, %vfalse : vector<42xi1>, vector<42xf32>

// Full vector selection.
%vx = std.select %cond, %vtrue, %vfalse : vector<42xf32>

pattern Select :: Location -> Type -> operand -> operand -> operand -> AbstractOperation operand

A pattern for std.select.

select :: MonadBlockBuilder m => Type -> Value -> Value -> Value -> m Value

A builder for std.select.

splat

Broadcast the operand to all elements of the result vector or tensor. The operand has to be of integerindexfloat type. When the result is a tensor, it has to be statically shaped.

Example:

%s = load %A[%i] : memref<128xf32>
%v = splat %s : vector<4xf32>
%t = splat %s : tensor<8x16xi32>

TODO: This operation is easy to extend to broadcast to dynamically shaped tensors in the same way dynamically shaped memrefs are handled.

// Broadcasts %s to a 2-d dynamically shaped tensor, with %m, %n binding
// to the sizes of the two dynamic dimensions.
%m = "foo"() : () -> (index)
%n = "bar"() : () -> (index)
%t = splat %s [%m, %n] : tensor<?x?xi32>

pattern Splat :: Location -> Type -> operand -> AbstractOperation operand

A pattern for std.splat.

splat :: MonadBlockBuilder m => Type -> Value -> m Value

A builder for std.splat.

switch

The switch terminator operation represents a switch on a signless integer value. If the flag matches one of the specified cases, then the corresponding destination is jumped to. If the flag does not match any of the cases, the default destination is jumped to. The count and types of operands must align with the arguments in the corresponding target blocks.

Example:

switch %flag : i32, [
  default: ^bb1(%a : i32),
  42: ^bb1(%b : i32),
  43: ^bb3(%c : i32)
]